修改了解包过程，使数据传输正确

+++ demo_arm/base/cv_comm/cv_comm.cpp
@@ -233,12 +233,12 @@ void CVComm::rxCallback(void) {
           //解包
           uint8_t tmp[sizeof(auto_exchange_pc2board_msg_)];
           memcpy(tmp,&auto_exchange_pc2board_msg_,rx_.frame.data_len);
-          auto_exchange_pc2board_msg_.TransX = ( tmp[0]|tmp[1]<<8 )/ 10.f;
-          auto_exchange_pc2board_msg_.TransY = ( tmp[2]|tmp[3]<<8 )/ 10.f;
-          auto_exchange_pc2board_msg_.TransZ = ( tmp[4]|tmp[5]<<8 )/ 10.f;
-          auto_exchange_pc2board_msg_.RotX = ( tmp[6]|tmp[7]<<8 )/ 10000.f;
-          auto_exchange_pc2board_msg_.RotY = ( tmp[8]|tmp[9]<<8 )/ 10000.f;
-          auto_exchange_pc2board_msg_.RotZ = ( tmp[10]|tmp[11]<<8 )/ 10000.f;
+          auto_exchange_pc2board_msg_.TransX = (int16_t)( tmp[0]|tmp[1]<<8 )/ 10.f;
+          auto_exchange_pc2board_msg_.TransY = (int16_t)( tmp[2]|tmp[3]<<8 )/ 10.f;
+          auto_exchange_pc2board_msg_.TransZ = (int16_t)( tmp[4]|tmp[5]<<8 )/ 10.f;
+          auto_exchange_pc2board_msg_.RotX = (int16_t)( tmp[6]|tmp[7]<<8 )/ 10000.f;
+          auto_exchange_pc2board_msg_.RotY = (int16_t)( tmp[8]|tmp[9]<<8 )/ 10000.f;
+          auto_exchange_pc2board_msg_.RotZ = (int16_t)( tmp[10]|tmp[11]<<8 )/ 10000.f;
           //解包完成
           auto_exchange_connect_.refresh();
       } else {
